fix(aero_realtime_omni): switch to sync-forward + inline code_predictor for wav
End-to-end pipeline works with 3 linked design pivots:

1. deploy yaml async_chunk: true -> false. Every stage-0 token step triggers
   orchestrator _forward_to_next_stage which submits a fresh stage-1 request.
   No need for resumable stage-1/2 requests or chunk-adapter lifecycle.

2. talker.forward now runs the 15-step residual code_predictor inline and
   returns OmniOutput(multimodal_outputs={"codes": {"audio": frame}}) with
   the [1, num_code_groups] codec frame. This ships codes to code2wav on the
   same forward step. talker.postprocess reduced to caching hidden_states.last;
   talker.make_omni_output reduced to an identity for OmniOutput passthrough.

3. code_predictor sampling knobs (do_sample, temperature, top_k, top_p) read
   from vllm_config.model_config.subtalker_sampling_params (yaml stage-1
   subtalker_sampling_params block) with training-time defaults as fallback.
   Same pattern as qwen3_tts talker.

thinker2talker (sync path) casts hidden/embed to float32 before serialization
because the wire serializer (numpy) does not support bfloat16; talker preprocess
casts back to bfloat16 for the trunk model.

Verified on 30 s Charades video with the reference checkpoint: 373 text tokens
+ 374 audio deltas -> 29.84 s wav with dynamic RMS variation (real speech
waveform, matches video duration).

+++ vllm_omni/model_executor/stage_input_processors/aero_realtime_omni.py
@@ -146,12 +146,15 @@ def thinker2talker(
 
         payload: OmniPayload = {
             "embed": {
-                "prefill": embed.detach().to("cpu")
+                # Serializer uses numpy which does not support bfloat16, so cast
+                # to float32 on the wire; talker preprocess casts back to
+                # bfloat16 for the trunk model.
+                "prefill": embed.detach().to("cpu", dtype=torch.float32)
                 if isinstance(embed, torch.Tensor)
                 else torch.empty(0)
             },
             "hidden_states": {
-                "output": hs.detach().to("cpu")
+                "output": hs.detach().to("cpu", dtype=torch.float32)
                 if isinstance(hs, torch.Tensor)
                 else torch.empty(0)
             },
